main page was splitted into three componenets
