+++ src/components/data-grid/data-grid-header.module.css
@@ -1,0 +1,14 @@
+.grid_container {
+  width: 100%;
+  display: grid;
+  grid-template-columns: repeat(6, 1fr);
+  padding: 5px;
+  background-color: #f0f0f0;
+  border-radius: 8px;
+}
+
+.grid_header {
+  padding: 5px;
+  justify-self: center;
+  align-self: center;
+}
